ENH turn pre-release into a full one when it works
